Add support for showing submitted date, from review
In addition to author and committer date, make it possible to
show a "submitted" date from when the commit was last submitted.
This is useful with reviewed branches from Gerrit Code Review,
where there might be some time between the commit and submit.

All information from Gerrit is stored in refs/notes/review,
where the "Submitted-by" and "Submitted-at" are the ones used.
In addition to some new %S tags for the pretty log format,
also add a new --fullest option for showing all three dates.

Signed-off-by: Anders Björklund <anders@itension.se>